Moved data logic from a world mixing into a DataEngine class. This helps reduce clutter in the @world namespace and makes replacing/testing components easier.
